requirements.txt
mysql-connector-python
bcrypt
pillow
python-dotenv
